Update fits2props.pro
current implementation defaults to FRIENDS=1 if either PHYSICAL flag not set, or both BMFRIENDS and FRIENDS are not set. This is in units of pixels, which means the search box for local maxima is only 3x3 in (x,y) by default - documentation states default is 3x3 beams. This change should give the functionality described by the documentation. My guess is that the 3x3 pixel box is unintended since interferometers typically oversample by ~5x and so the local maximum search was being done at sub-resolution. Many of the spurious peaks fail the subsequent decomposition tests, but this should speed things up and remove some pathological behavior. NOTE: needs some testing - I will do some and report back.